Include new system update system

+++ recipes-core/minke-run/run.sh
@@ -1,18 +1,22 @@
 #! /bin/sh
 
 # Install base images. Remove as we go so we don't do it again
-for image in /usr/share/minke/*.tar.gz; do
-  if [ "${image}" != "/usr/share/minke/*.tar.gz" ]; then
-    zcat ${image} | docker image load -q
-    rm -f ${image}
-  fi
-done
+if [ ! -f /usr/share/minke/docker-image-install-done ]; then
+  for image in /usr/share/minke/*.tar.gz; do
+    if [ "${image}" != "/usr/share/minke/*.tar.gz" ]; then
+      zcat ${image} | docker image load -q
+      rm -f ${image}
+    fi
+  done
+  touch /usr/share/minke/docker-image-install-done
+fi
 
 # Make sure various bind points exist
 mkdir -p /minke /minke/fs /minke/db /minke/skeletons/local
 touch /etc/timezone /etc/hostname /etc/systemd/network/bridge.network
 
 # Use the local nameserver
+OLDRESOLV=$(cat /etc/resolv.conf)
 echo "nameserver 127.0.0.1" > /etc/resolv.conf
 
 # Check that the Docker home network is still consistent with our IP and default route. If not, we delete the
@@ -26,11 +30,11 @@ if [ "${DOCKERIPNET}" != "${ORIGINALIPNET}" ]; then
 fi
 
 RESTART_REASON=/tmp/minke-restart-reason
-echo "reboot" > ${RESTART_REASON}
+echo "exit" > ${RESTART_REASON}
+TRACER_OUT=/tmp/tracer.out
+touch ${TRACER_OUT}
 
 while true; do 
-
-
   docker container rm minke
   docker run --name minke \
     --privileged \
@@ -39,17 +43,21 @@ while true; do
     -v /etc/hostname:/etc/hostname \
     -v /etc/systemd/network/bridge.network:/etc/systemd/network/bridge.network \
     -v ${RESTART_REASON}:${RESTART_REASON} \
+    -v ${TRACER_OUT}:${TRACER_OUT} \
     -v /var/run/docker.sock:/var/run/docker.sock \
     --mount type=bind,source=/minke/fs,target=/minke/fs,bind-propagation=rshared \
     --mount type=bind,source=/minke/db,target=/minke/db,bind-propagation=rshared \
     --mount type=bind,source=/minke/skeletons/local,target=/app/skeletons/local,bind-propagation=rshared \
     --network=host \
     registry.gitlab.com/minkebox/minke
-
   case "$(cat ${RESTART_REASON})" in
-    halt) shutdown -h now ;;
-    reboot) shutdown -r now ;;
+    halt)   systemctl poweroff ;;
+    reboot) systemctl reboot ;;
+    update) systemctl start dnf-automatic-restart ;;
+    exit)   exit ;;
     *) ;;
   esac
-
 done
+
+# Put back the old resolv
+echo ${OLDRESOLV} > /etc/resolv.conf
